Addd additional memory resource.
The seedfinder constructor was changed to accept addiitonal memory resource
which is meant to be device memory resource. By default, the pointer is null. In this case,
it is assumed that only shared memory resource was passed to the constructor.

The CreateSeedsForGroupSycl was changed to behave differently depending on the number of
memory resources that were passed to the function. For now the approach is the following:

1. Smart pointers are created
2. Inside the if-else scope, the pointers are assigned to newly created vectors with either device/host or shared memory resource.
3. Later parts of the code make use of the pointers.

After implementing this logic to the vectors of space points, on NVIDIA gpu the output gives segmentation fault.